+++ test/test_connectivity.jl
@@ -8,7 +8,7 @@ using NNlib: NNlib
 using Test
 
 const FIRST_ORDER_SET_TYPES = (
-    BitSet, Set{UInt64}, DuplicateVector{UInt64}, RecursiveSet{UInt64}, SortedVector{UInt64}
+    BitSet, Set{Int}, DuplicateVector{Int}, RecursiveSet{Int}, SortedVector{Int}
 )
 NNLIB_ACTIVATIONS_S = (
     NNlib.σ,
