+++ test/test_hessian.jl
@@ -6,7 +6,7 @@ using SpecialFunctions: erf, beta
 using Test
 
 const SECOND_ORDER_SET_TYPES = (
-    BitSet, Set{UInt64}, DuplicateVector{UInt64}, RecursiveSet{UInt64}, SortedVector{UInt64}
+    BitSet, Set{Int}, DuplicateVector{Int}, RecursiveSet{Int}, SortedVector{Int}
 )
 
 @testset "Global Hessian" begin
